+++ src/@tests/unit/infra/middlewares/cors.test.ts
@@ -23,6 +23,7 @@ Deno.test('Cors validation pipe', async () => {
     id: '',
     url: baseUrl,
     locals: {},
+    cookies: {},
   }
 
   assertThrows(